Move TAP code out of ETH-OC device, add USE flag
